SQLi-Lab/second-order+insert/requirements.txt
Flask==3.1.2
mysql-connector-python==9.6.0